changing the directory for index.js
