+++ server/src/set-up/env.js
@@ -1,5 +1,6 @@
 import { config } from "dotenv";
 config();
+export const PORT= process.env.PORT
 export const JWT_KEY = process.env.JWT_KEY
 export const NODE_ENV = process.env.NODE_ENV
 export const DATABASE_URL =
